+++ integration/util.go
@@ -242,3 +242,29 @@ func sortItems(items []map[string]types.AttributeValue) {
 		return titleI < titleJ
 	})
 }
+
+type TestContext struct {
+	ddbLocal *dynamodb.Client
+	baddb    *dynamodb.Client
+	shutdown func()
+}
+
+func setupTest(t *testing.T) *TestContext {
+	ddbLocal := newDdbLocalClient()
+	baddb := newBaddbClient()
+	cleanDdbLocal(ddbLocal)
+	shutdown := startServer()
+
+	_, ddbErr := createTable(ddbLocal)
+	_, baddbErr := createTable(baddb)
+	if ddbErr != nil || baddbErr != nil {
+		t.Fatalf("failed to create table: ddbErr=%v, baddbErr=%v", ddbErr, baddbErr)
+	}
+
+	return &TestContext{
+		ddbLocal: ddbLocal,
+		baddb:    baddb,
+		shutdown: shutdown,
+	}
+
+}